wrench act uses tool proc!

+++ code/game/machinery/alarm.dm
@@ -231,14 +231,14 @@
 		if(!get_danger_level(target_temperature, TLV["temperature"]) && abs(environment.temperature - target_temperature) > 2.0)
 			update_use_power(POWER_USE_ACTIVE)
 			regulating_temperature = 1
-			visible_message("\The [src] clicks as it starts [environment.temperature > target_temperature ? "cooling" : "heating"] the room.",\
+			visible_message("[src] clicks as it starts [environment.temperature > target_temperature ? "cooling" : "heating"] the room.",\
 			"You hear a click and a faint electronic hum.")
 	else
 		//check for when we should stop adjusting temperature
 		if (get_danger_level(target_temperature, TLV["temperature"]) || abs(environment.temperature - target_temperature) <= 0.5)
 			update_use_power(POWER_USE_IDLE)
 			regulating_temperature = 0
-			visible_message("\The [src] clicks quietly as it stops [environment.temperature > target_temperature ? "cooling" : "heating"] the room.",\
+			visible_message("[src] clicks quietly as it stops [environment.temperature > target_temperature ? "cooling" : "heating"] the room.",\
 			"You hear a click as a faint electronic humming stops.")
 
 	if (regulating_temperature)
@@ -649,7 +649,7 @@
 		return STATUS_CLOSE
 
 	if(aidisabled && issilicon(user))
-		to_chat(user, SPAN_WARNING("AI control for \the [src] interface has been disabled."))
+		to_chat(user, SPAN_WARNING("AI control for [src] interface has been disabled."))
 		return STATUS_CLOSE
 
 	. = shorted ? STATUS_DISABLED : STATUS_INTERACTIVE
@@ -837,19 +837,21 @@
 
 /obj/machinery/alarm/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
-	if(buildstage == 0)
-		to_chat(user, "You remove the fire alarm assembly from the wall!")
-		var/obj/item/frame/air_alarm/frame = new /obj/item/frame/air_alarm(get_turf(user))
-		transfer_fingerprints_to(frame)
-		playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
-		qdel(src)
+	if(buildstage != 0)
+		return
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
+	to_chat(user, "You remove the fire alarm assembly from the wall!")
+	var/obj/item/frame/air_alarm/frame = new /obj/item/frame/air_alarm(get_turf(user))
+	transfer_fingerprints_to(frame)
+	qdel(src)
 
 /obj/machinery/alarm/wirecutter_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
 	if(buildstage == 2)
 		if(!wiresexposed)
 			return
-		user.visible_message(SPAN_WARNING("[user] has cut the wires inside \the [src]!"), "You have cut the wires inside \the [src].")
+		user.visible_message(SPAN_WARNING("[user] has cut the wires inside [src]!"), "You have cut the wires inside [src].")
 		playsound(src.loc, 'sound/items/Wirecutter.ogg', 50, 1)
 		new/obj/item/stack/cable_coil(get_turf(src), 5)
 		buildstage = 1
@@ -873,12 +875,12 @@
 			if (isCoil(W))
 				var/obj/item/stack/cable_coil/C = W
 				if (C.use(5))
-					to_chat(user, SPAN_NOTICE("You wire \the [src]."))
+					to_chat(user, SPAN_NOTICE("You wire [src]."))
 					buildstage = 2
 					update_icon()
 					return TRUE
 				else
-					to_chat(user, SPAN_WARNING("You need 5 pieces of cable to do wire \the [src]."))
+					to_chat(user, SPAN_WARNING("You need 5 pieces of cable to do wire [src]."))
 					return TRUE
 
 		if(0)
@@ -1021,8 +1023,8 @@ FIRE ALARM
 	if(buildstage == 2)
 		detecting = !detecting
 		user.visible_message(
-			SPAN_NOTICE("\The [user] has [detecting? "re" : "dis"]connected \the [src]'s detecting unit!"),
-			SPAN_NOTICE("You have [detecting? "re" : "dis"]connected \the [src]'s detecting unit.")
+			SPAN_NOTICE("[user] has [detecting? "re" : "dis"]connected [src]'s detecting unit!"),
+			SPAN_NOTICE("You have [detecting? "re" : "dis"]connected [src]'s detecting unit.")
 		)
 
 /obj/machinery/firealarm/screwdriver_act(mob/living/user, obj/item/tool)
@@ -1036,18 +1038,20 @@ FIRE ALARM
 
 /obj/machinery/firealarm/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
-	if(buildstage == 0)
-		to_chat(user, "You remove the fire alarm assembly from the wall!")
-		new /obj/item/frame/fire_alarm(get_turf(user))
-		playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
-		qdel(src)
+	if(buildstage != 0)
+		return
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
+	to_chat(user, "You remove the fire alarm assembly from the wall!")
+	new /obj/item/frame/fire_alarm(get_turf(user))
+	qdel(src)
 
 /obj/machinery/firealarm/wirecutter_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
 	if(buildstage == 2)
 		user.visible_message(
-			SPAN_NOTICE("\The [user] has cut the wires inside \the [src]!"),
-			SPAN_NOTICE("You have cut the wires inside \the [src].")
+			SPAN_NOTICE("[user] has cut the wires inside [src]!"),
+			SPAN_NOTICE("You have cut the wires inside [src].")
 		)
 		new/obj/item/stack/cable_coil(get_turf(src), 5)
 		playsound(src.loc, 'sound/items/Wirecutter.ogg', 50, 1)
@@ -1063,12 +1067,12 @@ FIRE ALARM
 				if(istype(W, /obj/item/stack/cable_coil))
 					var/obj/item/stack/cable_coil/C = W
 					if (C.use(5))
-						to_chat(user, SPAN_NOTICE("You wire \the [src]."))
+						to_chat(user, SPAN_NOTICE("You wire [src]."))
 						buildstage = 2
 						update_icon()
 						return TRUE
 					else
-						to_chat(user, SPAN_WARNING("You need 5 pieces of cable to wire \the [src]."))
+						to_chat(user, SPAN_WARNING("You need 5 pieces of cable to wire [src]."))
 						return TRUE
 			if(0)
 				if(istype(W, /obj/item/firealarm_electronics))
@@ -1078,7 +1082,7 @@ FIRE ALARM
 					update_icon()
 					return TRUE
 
-	to_chat(user, SPAN_WARNING("You fumble with \the [W] and trigger the alarm!"))
+	to_chat(user, SPAN_WARNING("You fumble with [W] and trigger the alarm!"))
 	alarm()
 	return TRUE
