+++ code/game/objects/structures/drain.dm
@@ -13,12 +13,13 @@
 
 /obj/structure/hygiene/drain/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
 	var/obj/item/drain/drain_item = new(loc)
 	transfer_fingerprints_to(drain_item)
-	playsound(src, 'sound/items/Ratchet.ogg', 50, TRUE)
 	user.visible_message(
-		SPAN_NOTICE("\The [user] unwrenches \the [src] from the floor with \a [tool]."),
-		SPAN_NOTICE("You unwrench \the [src] from the floor with \the [tool].")
+		SPAN_NOTICE("[user] unwrenches [src] from the floor with [tool]."),
+		SPAN_NOTICE("You unwrench [src] from the floor with [tool].")
 	)
 	qdel(src)
 
@@ -30,8 +31,8 @@
 		return
 	welded = !welded
 	user.visible_message(
-		SPAN_NOTICE("\The [user] [welded ? "un" : "welds"] \the [src] with \a [tool]."),
-		SPAN_NOTICE("You [welded ? "un" : "weld"] \the [src] with \the [tool].")
+		SPAN_NOTICE("[user] [welded ? "un" : "welds"] [src] with [tool]."),
+		SPAN_NOTICE("You [welded ? "un" : "weld"] [src] with [tool].")
 	)
 	update_icon()
 
@@ -60,10 +61,11 @@
 /obj/item/drain/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
 	if (!isturf(loc))
-		USE_FEEDBACK_FAILURE("\The [src] needs to be placed on the floor before you can secure it.")
+		USE_FEEDBACK_FAILURE("[src] needs to be placed on the floor before you can secure it.")
+		return
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
 		return
 	new constructed_type(src.loc)
-	playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
 	to_chat(user, SPAN_WARNING("[user] wrenches the [src] down."))
 	qdel(src)
 
@@ -77,7 +79,7 @@
 	. = ..()
 	if(!welded)
 		closed = !closed
-		user.visible_message(SPAN_NOTICE("\The [user] has [closed ? "closed" : "opened"] the drain."))
+		user.visible_message(SPAN_NOTICE("[user] has [closed ? "closed" : "opened"] the drain."))
 	update_icon()
 
 /obj/structure/hygiene/drain/bath/on_update_icon()
