+++ code/modules/assembly/shock_kit.dm
@@ -24,14 +24,17 @@
 
 /obj/item/assembly/shock_kit/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
-	if(!status)
-		part1.dropInto(loc)
-		part2.dropInto(loc)
-		part1.master = null
-		part2.master = null
-		part1 = null
-		part2 = null
-		qdel(src)
+	if(status)
+		return
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
+	part1.dropInto(loc)
+	part2.dropInto(loc)
+	part1.master = null
+	part2.master = null
+	part1 = null
+	part2 = null
+	qdel(src)
 
 /obj/item/assembly/shock_kit/attack_self(mob/user as mob)
 	part1.attack_self(user, status)
