+++ code/modules/atmospherics/components/unary/outlet_injector.dm
@@ -204,6 +204,8 @@
 
 /obj/machinery/atmospherics/unary/outlet_injector/wrench_act(mob/living/user, obj/item/tool)
 	. = ITEM_INTERACT_SUCCESS
+	if(!tool.use_as_tool(src, user, volume = 50, do_flags = DO_REPAIR_CONSTRUCT))
+		return
 	new /obj/item/pipe(loc, src)
 	qdel(src)
